EduTeX_Agents/requirements.txt
# EduTeX Agent API — Python Dependencies
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
pydantic>=2.10.0
python-multipart>=0.0.18

# AI / LLM
google-genai>=1.0.0
openai>=1.60.0

# Math CAS
sympy>=1.13.0

# Utilities
python-dotenv>=1.0.0
